+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Support for new version 3.7.
-- New method `request_welcome_message` in customer-api v3.7.
+- New method `request_welcome_message` in customer-api v3.6 and v3.7.
 - New method `delete_event` in customer-api v3.6.
 - New methods in configuration-api v3.6 for greetings: `create_greeting`, `delete_greeting`, `get_greeting`, `update_greeting`, `list_greetings`.
 - New methods in agent-api v3.6: `send_thinking_indicator`, `send_message_preview`.